requirements.txt
openai==1.74.0
openai_whisper==20240930
torch==2.6.0
tqdm==4.67.1
